% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.ontology.col.R
\name{add.ontology.col}
\alias{add.ontology.col}
\title{Add extra attributes to the datumTriples to aid plotting}
\usage{
add.ontology.col(
  triples,
  triples.to.find,
  datumAttributeCol = "plot.core",
  datumValueCol = "Ontology",
  datumValueNot = "XML"
)
}
\arguments{
\item{triples}{dataframe of datumTriples}

\item{triples.to.find}{List of keywords to mark entries.}

\item{datumAttributeCol}{Name of the attribute column}

\item{datumValueCol}{Value to use for selected rows}

\item{datumValueNot}{Value to use for rows not selected}
}
\value{
dataframe of datumTriples
}
\description{
Add extra attributes to the datumTriples to aid plotting
}
