% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.triples.to.network.R
\name{advanced.triples.to.network}
\alias{advanced.triples.to.network}
\title{Advance Triple Network}
\usage{
advanced.triples.to.network(x)
}
\arguments{
\item{x}{Set of datum triples}
}
\value{
Network object
}
\description{
This function has enhanced checking for missing newtwrk nodes
}
