% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.as.triple.R
\name{data.as.triple}
\alias{data.as.triple}
\title{Turn Data Into Datum Triple format}
\usage{
data.as.triple(x)
}
\arguments{
\item{x}{Data.frame with datumAttribute and named columns with values}
}
\value{
A data.frame of triples datumEntity, datumAttribute, datumValue.
}
\description{
A convenience function for turning data.frames into long datum Triple format.
}
