% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.datum.parent.R
\name{get.datum.parent}
\alias{get.datum.parent}
\title{Datum Entity Parent}
\usage{
get.datum.parent(x)
}
\arguments{
\item{x}{A syntactically correct XML file.}
}
\value{
A syntactically correct XML file with datumParent set.
}
\description{
Infers the datumEntity of the parent node. Used to help function set.datum.parent
}
