% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tag.value.pairs.R
\name{get.tag.value.pairs}
\alias{get.tag.value.pairs}
\title{Get Tag Value Pairs}
\usage{
get.tag.value.pairs(x, drop.tag.cols = TRUE)
}
\arguments{
\item{x}{Datum Triple format data.}

\item{drop.tag.cols}{Logical. If TRUE drop the columns 'tag' and 'value'.}
}
\value{
datumTriples with tag value pairs.
}
\description{
Expand tag value pairs in the the data into datumAttribute and datumValue.
This is a specific method of encoding data found in EA files.
}
