% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.xml.attrs.R
\name{get.xml.attrs}
\alias{get.xml.attrs}
\title{Get XML Attributes}
\usage{
get.xml.attrs(xml.object)
}
\arguments{
\item{xml.object}{An XML file}
}
\value{
A list of lists: a sorted list; the number of attributes; a
paragraph.
}
\description{
A convience function to list all the attributes in an XML file.
}
