% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.datum.entity.R
\name{infer.datum.entity}
\alias{infer.datum.entity}
\title{Infer Datum Entity}
\usage{
infer.datum.entity(x, internal.id = "id")
}
\arguments{
\item{x}{A syntactically correct XML file.}

\item{internal.id}{This is the XML attribute name used for unique ID. This
default value is associated with Enterprise Architect.}
}
\value{
An XML file with datum Entity set.
}
\description{
Checks to see if a system generated unique id exists. If it does, this is
used as a datumEntity, if not, one is randomly generated.
}
