% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.graph.ml.R
\name{save.graph.ml}
\alias{save.graph.ml}
\title{Save Graph In GML Format}
\usage{
save.graph.ml(x, name.prefix = "subgraph")
}
\arguments{
\item{x}{Graph object}

\item{name.prefix}{Name prefix for the file}
}
\value{
x unchanged and saves file.
}
\description{
Saves a graph object as GMl so it can be read by Gephi. The form .ml is left
over from an earlier function and left for compatibility.
}
