% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.datum.entity.R
\name{set.datum.entity}
\alias{set.datum.entity}
\title{Set Datum Entity}
\usage{
set.datum.entity(x)
}
\arguments{
\item{x}{A syntactically correct XML file.}
}
\value{
A syntactically correct XML file with unique datumEntity attributes
}
\description{
Sets a unique datumEntity value for each element of an XML file.
}
