% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.as.wide.R
\name{triple.as.wide}
\alias{triple.as.wide}
\title{Triple As Wide}
\usage{
triple.as.wide(x)
}
\arguments{
\item{x}{Three column datum Triple format data.}
}
\value{
Wide data.frame.
}
\description{
Expands datumTriples into a data.frame with one row per datumEntity and on
column per datumAttribue datumValue pair.
}
