% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.clean.dups.R
\name{triple.clean.dups}
\alias{triple.clean.dups}
\title{Triple Clean Duplicates}
\usage{
triple.clean.dups(x)
}
\arguments{
\item{x}{Set of triples}
}
\value{
Set of triples with duplicates with new datumEntity values
}
\description{
Triple Clean Duplicates
}
