% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples.to.context.network.R
\name{triples.to.context.network}
\alias{triples.to.context.network}
\title{Triples To Context Network}
\usage{
triples.to.context.network(
  x,
  cols.to.include = c("id", "datumEntity", "tag.type", "nominal.name"),
  drop.xml.edges = TRUE,
  correct.dup.atts = TRUE
)
}
\arguments{
\item{x}{Set of datum triples}

\item{cols.to.include}{List of columns to include.}

\item{drop.xml.edges}{Logical. Should we drop the xml edges}

\item{correct.dup.atts}{combine duplicate attributes}
}
\value{
Network object
}
\description{
Converts triples into a network.
}
