% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples.to.network.R
\name{triples.to.network}
\alias{triples.to.network}
\title{Triples To Network}
\usage{
triples.to.network(x)
}
\arguments{
\item{x}{Datum Triples of an XML document}
}
\value{
A ggplot2 Network data.frame.
}
\description{
Converts datum Triples of an XML document into a gglot2 network plot with
tag.type as the node and descendent as the edges. This gives some insight
into the document structure.
}
