% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.attrs.to.dataframe.R
\name{xml.attrs.to.dataframe}
\alias{xml.attrs.to.dataframe}
\title{XML Attributes To Data Frame}
\usage{
xml.attrs.to.dataframe(x)
}
\arguments{
\item{x}{An XML node.}
}
\value{
A data frame containing the XML attributes.
}
\description{
A convenience function to extract XML attributes from a node and return them
as a data.frame.
}
