% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.triples.dataset..R
\docType{data}
\name{xml.triples.dataset}
\alias{xml.triples.dataset}
\title{XML Triples Dataset.}
\format{
\subsection{\verb{datum Triples}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{datumEntity}{A unique identifier}
\item{datumAtribute}{Attribute name}
\item{datumValue}{Value of attribute}
...
}
}
}
\source{
Created as an example.
}
\usage{
xml.triples.dataset
}
\description{
A sample XML file created with Enterprise Architect software
}
\keyword{datasets}
