% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.new.nodes.R
\name{any.new.nodes}
\alias{any.new.nodes}
\title{Any New Nodes}
\usage{
\method{any}{new.nodes}(new.nodes, existing.nodes, new.node.type = "Orphan")
}
\arguments{
\item{new.nodes}{List of nodes}

\item{existing.nodes}{List of nodes.}

\item{new.node.type}{Name for any new nodes.}
}
\value{
List of nodes
}
\description{
An internal function used to help build igraph networks by finding nodes not
included in derived lists and labelling them. This is useful when
constructing networks with duplicate IDs and highlighting orphan links.
}
