% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.datum.triple.R
\name{make.datum.triple}
\alias{make.datum.triple}
\title{Make Datum Triple}
\usage{
make.datum.triple(x)
}
\arguments{
\item{x}{A syntactically correct XML file with datumEntity and other
attributes set.}
}
\value{
A data.frame of triples datumEntity, datumAttribute, datumValue.
}
\description{
Extracts attributes from a XML element and presents them as a datum Triple in
the form of a three column data frame.
}
