% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.file.to.triples.R
\name{xml.file.to.triples}
\alias{xml.file.to.triples}
\title{XML File To Triples}
\usage{
xml.file.to.triples(
  the.file.path = the.file.path,
  file.number = 1,
  xpath = "/*"
)
}
\arguments{
\item{the.file.path}{The path to the file directory}

\item{file.number}{A number - which file to use if there are more than 1.}

\item{xpath}{The Xpath the default is root}
}
\value{
List of three values: triples; file.name; full.file.path.
}
\description{
This reads a single XML file from a path and turns it into datumTriples.
There is some resilience to the presence of duplicate IDs. This has only been
tested on files generated by Enterprise Architect.
}
